feat: refatoração no padrão SOLID e DDD
